Move mutual info functions to statistics_utils file (#584)
* move over mi helpers

* move over get mode

* move over mutual info dict function

* Cleanup

* Add release note

* Fix after rebase

* use method for mutual information
