+++ woodwork/datatable.py
@@ -10,6 +10,7 @@ from woodwork.datacolumn import DataColumn
 from woodwork.exceptions import ColumnNameMismatchWarning
 from woodwork.indexers import _iLocIndexer
 from woodwork.logical_types import Boolean, Datetime, Double, LatLong
+from woodwork.statistics_utils import _get_mode
 from woodwork.type_sys.utils import (
     _get_ltype_class,
     _is_numeric_series,
@@ -17,7 +18,6 @@ from woodwork.type_sys.utils import (
 )
 from woodwork.utils import (
     _convert_input_to_set,
-    _get_mode,
     _new_dt_including,
     get_valid_mi_types,
     import_or_none,
