+++ woodwork/statistics_utils.py
@@ -1,6 +1,8 @@
 import numpy as np
+import pandas as pd
+from sklearn.metrics.cluster import normalized_mutual_info_score
 
-from woodwork.logical_types import Datetime, LatLong
+from woodwork.logical_types import Datetime, Double, LatLong
 from woodwork.schema_column import (
     _is_col_boolean,
     _is_col_categorical,
@@ -8,7 +10,7 @@ from woodwork.schema_column import (
     _is_col_numeric
 )
 from woodwork.type_sys.utils import _get_ltype_class
-from woodwork.utils import _get_mode, import_or_none
+from woodwork.utils import get_valid_mi_types, import_or_none
 
 dd = import_or_none('dask.dataframe')
 ks = import_or_none('databricks.koalas')
@@ -98,3 +100,135 @@ def _get_describe_dict(dataframe, include=None):
         values["semantic_tags"] = semantic_tags
         results[column_name] = values
     return results
+
+
+def _get_mode(series):
+    """Get the mode value for a series"""
+    mode_values = series.mode()
+    if len(mode_values) > 0:
+        return mode_values[0]
+    return None
+
+
+def _replace_nans_for_mutual_info(schema, data):
+    """
+    Replace NaN values in the dataframe so that mutual information can be calculated
+
+    Args:
+        schema (woodwork.Schema): Woodwork typing info for the data
+        data (pd.DataFrame): dataframe to use for calculating mutual information
+
+    Returns:
+        pd.DataFrame: data with nans replaced with either mean or mode
+
+    """
+    # replace or remove null values
+    for column_name in data.columns[data.isnull().any()]:
+        column = schema.columns[column_name]
+        series = data[column_name]
+
+        if _is_col_numeric(column) or _is_col_datetime(column):
+            mean = series.mean()
+            if isinstance(mean, float) and not _get_ltype_class(column['logical_type']) == Double:
+                data[column_name] = series.astype('float')
+            data[column_name] = series.fillna(mean)
+        elif _is_col_categorical(column) or _is_col_boolean(column):
+            mode = _get_mode(series)
+            data[column_name] = series.fillna(mode)
+    return data
+
+
+def _make_categorical_for_mutual_info(schema, data, num_bins):
+    """Transforms dataframe columns into numeric categories so that
+    mutual information can be calculated
+
+    Args:
+        schema (woodwork.Schema): Woodwork typing info for the data
+        data (pd.DataFrame): dataframe to use for caculating mutual information
+        num_bins (int): Determines number of bins to use for converting
+            numeric features into categorical.
+
+    Returns:
+        pd.DataFrame: data with values transformed and binned into numeric categorical values
+    """
+
+    for col_name in data.columns:
+        column = schema.columns[col_name]
+        if _is_col_numeric(column):
+            # bin numeric features to make categories
+            data[col_name] = pd.qcut(data[col_name], num_bins, duplicates="drop")
+        # Convert Datetimes to total seconds - an integer - and bin
+        if _is_col_datetime(column):
+            data[col_name] = pd.qcut(data[col_name].astype('int64'), num_bins, duplicates="drop")
+        # convert categories to integers
+        new_col = data[col_name]
+        if str(new_col.dtype) != 'category':
+            new_col = new_col.astype('category')
+        data[col_name] = new_col.cat.codes
+    return data
+
+
+def _get_mutual_information_dict(dataframe, num_bins=10, nrows=None):
+    """
+    Calculates mutual information between all pairs of columns in the DataFrame that
+    support mutual information. Logical Types that support mutual information are as
+    follows:  Boolean, Categorical, CountryCode, Datetime, Double, Integer, Ordinal,
+    SubRegionCode, and ZIPCode
+
+    Args:
+        dataframe (pd.DataFrame): Data containing Woodwork typing information
+            from which to calculate mutual information.
+        num_bins (int): Determines number of bins to use for converting
+            numeric features into categorical.
+        nrows (int): The number of rows to sample for when determining mutual info.
+        If specified, samples the desired number of rows from the data.
+            Defaults to using all rows.
+
+    Returns:
+        list(dict): A list containing dictionaries that have keys `column_1`,
+        `column_2`, and `mutual_info` that is sorted in decending order by mutual info.
+        Mutual information values are between 0 (no mutual information) and 1
+        (perfect dependency).
+        """
+    valid_types = get_valid_mi_types()
+    valid_columns = [col_name for col_name, col in dataframe.ww.columns.items() if (
+        col_name != dataframe.ww.index and _get_ltype_class(col['logical_type']) in valid_types)]
+
+    data = dataframe[valid_columns]
+    if dd and isinstance(data, dd.DataFrame):
+        data = data.compute()
+    if ks and isinstance(dataframe, ks.DataFrame):
+        data = data.to_pandas()
+
+    # cut off data if necessary
+    if nrows is not None and nrows < data.shape[0]:
+        data = data.sample(nrows)
+
+    # remove fully null columns
+    not_null_cols = data.columns[data.notnull().any()]
+    if set(not_null_cols) != set(valid_columns):
+        data = data.loc[:, not_null_cols]
+    # remove columns that are unique
+    not_unique_cols = [col for col in data.columns if not data[col].is_unique]
+    if set(not_unique_cols) != set(valid_columns):
+        data = data.loc[:, not_unique_cols]
+
+    data = _replace_nans_for_mutual_info(dataframe.ww.schema, data)
+    data = _make_categorical_for_mutual_info(dataframe.ww.schema, data, num_bins)
+
+    # calculate mutual info for all pairs of columns
+    mutual_info = []
+    col_names = data.columns.to_list()
+    for i, a_col in enumerate(col_names):
+        for j in range(i, len(col_names)):
+            b_col = col_names[j]
+            if a_col == b_col:
+                # Ignore because the mutual info for a column with itself will always be 1
+                continue
+            else:
+                mi_score = normalized_mutual_info_score(data[a_col], data[b_col])
+                mutual_info.append(
+                    {"column_1": a_col, "column_2": b_col, "mutual_info": mi_score}
+                )
+    mutual_info.sort(key=lambda mi: mi['mutual_info'], reverse=True)
+    return mutual_info
