+++ woodwork/tests/test_utils.py
@@ -29,7 +29,6 @@ from woodwork.type_sys.utils import (
 from woodwork.utils import (
     _convert_input_to_set,
     _get_column_logical_type,
-    _get_mode,
     _is_null_latlong,
     _is_s3,
     _is_url,
@@ -134,27 +133,6 @@ def test_list_semantic_tags():
         if name not in ['index', 'time_index', 'date_of_birth']:
             for log_type in log_type_list:
                 assert name in log_type.standard_tags
-
-
-def test_get_mode():
-    series_list = [
-        pd.Series([1, 2, 3, 4, 2, 2, 3]),
-        pd.Series(['a', 'b', 'b', 'c', 'b']),
-        pd.Series([3, 2, 3, 2]),
-        pd.Series([np.nan, np.nan, np.nan]),
-        pd.Series([pd.NA, pd.NA, pd.NA]),
-        pd.Series([1, 2, np.nan, 2, np.nan, 3, 2]),
-        pd.Series([1, 2, pd.NA, 2, pd.NA, 3, 2])
-    ]
-
-    answer_list = [2, 'b', 2, None, None, 2, 2]
-
-    for series, answer in zip(series_list, answer_list):
-        mode = _get_mode(series)
-        if answer is None:
-            assert mode is None
-        else:
-            assert mode == answer
 
 
 def test_read_csv_no_params(sample_df_pandas, tmpdir):
