+++ woodwork/utils.py
@@ -50,14 +50,6 @@ def _convert_input_to_set(semantic_tags, error_language='semantic_tags'):
     return semantic_tags
 
 
-def _get_mode(series):
-    """Get the mode value for a series"""
-    mode_values = series.mode()
-    if len(mode_values) > 0:
-        return mode_values[0]
-    return None
-
-
 def read_csv(filepath=None,
              name=None,
              index=None,
